Seperated stack out into its own file, renamed data structures
