+++ src/Parser/parser.hpp
@@ -1,4 +1,3 @@
-#include "../Data Structures/data_structures.hpp"
 #include "../Tokeniser/token.hpp"
 #include "tokenStreamIterator.hpp"
 #include "tree.hpp"
